+++ tests/routing/test_routing_get_bundle_sort_ports.py
@@ -29,7 +29,9 @@ def test_get_bundle_sort_ports(
         for i in range(N)
     ]
     left_ports.reverse()
-    routes = gf.routing.get_bundle(right_ports, left_ports)
+    routes = gf.routing.get_bundle(
+        right_ports, left_ports, sort_ports=True, enforce_port_ordering=False
+    )
 
     for i, route in enumerate(routes):
         c.add(route.references)
