+++ tests/test_component_from_yaml.py
@@ -314,7 +314,7 @@ placements:
 routes:
     optical:
         links:
-            left,o:1:3: right,o:1:3
+            left,o:1:3: right,o:3:1
 """
 
 sample_regex_connections_backwards = """
@@ -340,13 +340,14 @@ placements:
 routes:
     optical:
         links:
-            left,o:3:1: right,o:3:1
+            left,o:3:1: right,o:1:3
 """
 
 
 def test_connections_regex() -> None:
     c = from_yaml(sample_regex_connections)
-    route_names = ["left,o1:right,o1", "left,o2:right,o2", "left,o3:right,o3"]
+    route_names = ["left,o1:right,o3", "left,o2:right,o2", "left,o3:right,o1"]
+    c.show()
 
     length = 12.0
     for route_name in route_names:
@@ -355,7 +356,7 @@ def test_connections_regex() -> None:
 
 def test_connections_regex_backwargs() -> None:
     c = from_yaml(sample_regex_connections_backwards)
-    route_names = ["left,o1:right,o1", "left,o2:right,o2", "left,o3:right,o3"]
+    route_names = ["left,o3:right,o1", "left,o2:right,o2", "left,o1:right,o3"]
 
     length = 12.0
     for route_name in route_names:
